backend/requirements.txt
# ─── Django Core ─────────────────────────────────────────────────────────────
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0

# ─── Package Tools ─────────────────────────────────────────────────────────────
setuptools>=65.0

# ─── Database ─────────────────────────────────────────────────────────────────
# psycopg2-binary: PostgreSQL adapter (used on Render / production)
psycopg2-binary>=2.9,<3.0
# dj-database-url: parse DATABASE_URL env variable into Django DATABASES dict
dj-database-url>=2.1,<3.0

# ─── CORS ─────────────────────────────────────────────────────────────────────
django-cors-headers>=4.3,<5.0

# ─── Image Handling ───────────────────────────────────────────────────────────
Pillow>=10.0,<11.0

# ─── Environment Variables ────────────────────────────────────────────────────
python-dotenv>=1.0,<2.0

# ─── Production Server ────────────────────────────────────────────────────────
gunicorn>=21.0,<22.0

# ─── Static Files (production) ────────────────────────────────────────────────
whitenoise>=6.6,<7.0

# ─── Payment Gateway (Cashfree) ───────────────────────────────────────────────
# requests is used to call Cashfree REST API directly (no official Python SDK needed)
requests>=2.31,<3.0

# ─── Cloud Storage (Cloudinary — persistent media on Render) ─────────────────
# cloudinary: Official Cloudinary Python SDK
cloudinary>=1.36,<2.0
# django-cloudinary-storage: Django storage backend for Cloudinary
django-cloudinary-storage>=0.3,<0.4
